% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_wf.R
\name{gg_wf}
\alias{gg_wf}
\title{Waterfall Plots}
\usage{
gg_wf(
  y,
  data,
  group = NULL,
  panel = NULL,
  point_geom = c("bar", "point", "line"),
  point_color = "black",
  point_alpha = 1,
  point_size = 1,
  hline = 0,
  line_color = "black",
  line_type = "solid",
  line_alpha = 1,
  line_size = 1
)
}
\arguments{
\item{y}{Column to be plotted on  y axis (should be numeric).}

\item{data}{A data frame.}

\item{group}{A grouping factor for plotting different colors/dodge on the same axis.}

\item{panel}{A factor by which to facet the plot.}

\item{point_geom}{The type of geom to represent the data. Choices include "bar" (default), "point", or "line".}

\item{point_size, point_alpha, point_color}{Change parameters for point data. Size does not work for when geom is "bar".}

\item{hline}{The position of the horizontal line on the y-axis (default = 0).}

\item{line_size, line_color, line_alpha, line_type}{Change line parameters.}
}
\value{
A ggplot object
}
\description{
Function for waterfall differences visualizations.
Something to make effects from paired (pre-post) intervention studies.
}
\details{
Creates a simple waterfall plot.
}
\references{
Gillespie T. W. (2012). Understanding waterfall plots. Journal of the advanced practitioner in oncology, 3(2), 106–111. https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4093310/

Shao et al. (2014) Use and Misuse of Waterfall Plots, JNCI: Journal of the National Cancer Institute, 106(12), 331. https://doi.org/10.1093/jnci/dju331
}
