% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rom_test.R
\name{rom_test}
\alias{rom_test}
\title{Ratio of Means Test}
\usage{
rom_test(
  x,
  y = NULL,
  data = NULL,
  null = 1,
  ci = 0.95,
  ci_method = c("t", "normal"),
  alternative = c("two.sided", "less", "greater"),
  bias_c = TRUE,
  paired = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric or ordered vector, or a character name of one in `data`.
Any missing values (`NA`s) are dropped from the resulting vector. `x` can
also be a formula (see [`stats::wilcox.test()`]), in which case `y` is
ignored.}

\item{y}{A numeric or ordered vector, or a character name of one in `data`.
Any missing values (`NA`s) are dropped from the resulting vector. `x` can
also be a formula (see [`stats::wilcox.test()`]), in which case `y` is
ignored.}

\item{data}{An optional data frame containing the variables.}

\item{null}{a number indicating the value of the null hypothesis. See [stats::wilcox.test].}

\item{ci}{Confidence level of the interval.}

\item{ci_method}{Method for calculating confidence levels.}

\item{alternative}{a character string specifying the alternative hypothesis;
Controls the type of CI returned: `"two.sided"` (default, two-sided CI),
`"greater"` or `"less"` (one-sided CI). Partial matching is allowed (e.g.,
`"g"`, `"l"`, `"two"`...).}

\item{bias_c}{Bias correction for small samples. Default is TRUE.}

\item{paired}{If `TRUE`, the values of `x` and `y` are considered as paired.
This produces an effect size that is equivalent to the one-sample effect
size on `x - y`.}

\item{verbose}{Toggle warning messages on or off.}

\item{...}{Arguments passed to or from other methods. When `x` is a formula,
these can be `subset` and `na.action`.}
}
\value{
An object of the class `htest`.
\describe{
    \item{statistic}{The test statistic (z or t).}
    \item{estimate}{The ratio of means or "response rato" (RR).}
    \item{p.value}{The p-value for the test.}
    \item{null.value}{The specified hypothesized value of the RR.}
    \item{alternative}{A character string describing the alternative hypothesis.}
    \item{method}{The name of the method to report.}
    \item{data.name}{The title of the data.}
    \item{conf.int}{The confience interval for the RR.}
    \item{stderr}{The standard error of the RR.}

}
}
\description{
Performs a test on a ratio of means.
}
\details{
Ratio of means... To be added
vtype details
}
\examples{
# Use the sleep data

rom_test(mpg ~ am, data = mtcars, paired = FALSE)

}
\references{
Lajeunesse, M. J. (2011). On the meta‐analysis of response ratios for studies with correlated and multi‐group designs. Ecology, 92(11), 2049-2055 .https://doi.org/10.1890/11-0423.1

Lajeunesse, M. J. (2015). Bias and correction for the log response ratio in ecological meta‐analysis. Ecology, 96(8), 2056-2063. https://doi.org/10.1890/14-2402.1

Kieser, M. and Hauschke, D. (1999) Approximate Sample Sizes for Testing Hypotheses about the Ratio and Difference of Two Means. Journal of Biopharmaceutical Studies, 9(4)  641-650.

Hauschke, D. et al  (1999). Sample Size Determination for Proving Equivalence Based on the Ratio of Two Means for Normally Distributed Data. Statistics in Medicine, 18,  93-105.
}
