% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_sum.R
\name{gg_sum}
\alias{gg_sum}
\title{Summary plots}
\usage{
gg_sum(
  data,
  x,
  y,
  group = NULL,
  panel = NULL,
  trace = TRUE,
  sum_stat = c("mean", "median"),
  show_points = TRUE,
  show_summary = TRUE,
  show_slab = FALSE,
  err_width = 0.2,
  sum_size = 1.25,
  sum_alpha = 1,
  point_size = 1,
  point_alpha = 1,
  sum_color = "darkgrey",
  point_color = "black"
)
}
\arguments{
\item{data}{A data frame.}

\item{x, y}{Columns to be plotted on x or y axis (x should be a factor).}

\item{group}{A grouping factor for plotting different colors/dodge on the same axis.}

\item{panel}{A factor by which to facet the plot.}

\item{trace}{TRUE/FALSE to connect summaries with line.}

\item{sum_stat}{Type of summary statistic. Options include mean (defualts) or median. Error bars will be the SD or IQR (25th and 75th percentile) for mean and median, respectively.}

\item{show_points, show_summary, show_slab}{Logical indicator of what to show on plot.}

\item{err_width}{Width of error bars, or separation of groups.}

\item{sum_size, sum_alpha}{Change parameters for summary statistics.}

\item{point_size, point_alpha}{Change parameters for point data.}

\item{sum_color}{Color for summary statistic (only used with groups == NULL).}

\item{point_color}{Color for data points (only used with groups == NULL).}
}
\value{
A ggplot object.
}
\description{
Function for simple factorial experiment plots.
Something that should save time for plotting data.
}
\details{
Creats simple plots for summarizing factorial designs. See vignettes for examples.
}
