% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_sum.R
\name{gg_sum}
\alias{gg_sum}
\title{Summary plots}
\usage{
gg_sum(
  data,
  x,
  y,
  group = NULL,
  panel = NULL,
  trace = TRUE,
  show_points = TRUE,
  err_width = 0.2,
  sum_size = 1.25,
  sum_alpha = 1,
  point_size = 1,
  point_alpha = 1
)
}
\arguments{
\item{x, y}{Columns to be plotted on x or y axis (x should be a factor).}

\item{group}{A grouping factor for plotting different colors/dodge on the same axis.}

\item{panel}{A factor by which to facet the plot.}

\item{trace}{TRUE/FALSE to connect summaries with line.}

\item{show_points}{Logical indicator to show individual data points.}

\item{err_width}{Width of error bars, or separation of groups.}

\item{sum_size, sum_alpha}{Change parameters for summary statistics.}

\item{point_size, point_alpha}{Change parameters for point data.}
}
\value{
An ggplot object
}
\description{
Function for simple factorial experiment plots.
Something that should save time for plotting data.
}
\details{
Creats simple plots for summarizing factorial designs.
}
